@@ -92,40 +92,40 @@
     "tree-kill": "1.2.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.10",
-    "@angular/animations": "7.2.15",
-    "@angular/cdk": "7.3.7",
-    "@angular/cli": "7.3.9",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/compiler-cli": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular/flex-layout": "7.0.0-beta.24",
-    "@angular/forms": "7.2.15",
-    "@angular/language-service": "7.2.15",
-    "@angular/material": "7.3.7",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "8.2.3",
+    "@angular/cli": "8.3.29",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/flex-layout": "8.0.0-beta.27",
+    "@angular/forms": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@angular/material": "8.2.3",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@angularclass/hmr": "2.1.3",
     "@fortawesome/fontawesome-free": "5.9.0",
-    "@nebular/bootstrap": "3.5.0",
-    "@nebular/theme": "https://github.com/dragonbane0/nebular/tarball/2ab1e12d9c10b26bcc0680acde44048d63345991",
-    "@ng-bootstrap/ng-bootstrap": "4.2.1",
+    "@nebular/eva-icons": "4.6.0",
+    "@nebular/theme": "4.6.0",
     "@types/electron": "1.6.10",
     "@types/jasmine": "2.8.16",
     "@types/jasminewd2": "2.0.6",
     "@types/node": "8.9.5",
     "angular-dual-listbox": "4.7.0",
     "angular-file": "1.3.2",
-    "bootstrap": "4.0.0",
-    "codelyzer": "4.5.0",
-    "core-js": "2.6.9",
+    "browserslist": "^4.21.4",
+    "codelyzer": "^5.0.1",
+    "core-js": "3.25.1",
     "crc": "3.8.0",
     "electron": "4.2.6",
+    "eva-icons": "1.1.3",
     "hammerjs": "2.0.8",
     "intl": "1.2.5",
-    "ionicons": "2.0.1",
+    "ionicons": "4.6.3",
     "jasmine-core": "2.99.1",
     "jasmine-spec-reporter": "4.2.1",
     "jquery": "3.4.1",
@@ -134,15 +134,14 @@
     "karma-coverage-istanbul-reporter": "2.0.5",
     "karma-jasmine": "1.1.2",
     "karma-jasmine-html-reporter": "0.2.2",
-    "nebular-icons": "1.1.0",
     "ngx-color-picker": "https://github.com/dragonbane0/ngx-color-picker/tarball/7471e8595e5bf508813c5f6b1964e53ac9902448",
     "node-sass": "4.14.1",
     "normalize.css": "6.0.0",
     "pace-progress": "https://github.com/dragonbane0/pace/tarball/bcbd25563bf6193b223f856dc1668c0705fd8710",
-    "popper.js": "1.15.0",
+    "popper.js": "1.16.1",
     "protractor": "5.4.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "6.3.3",
+    "rxjs": "6.6.7",
     "rxjs-compat": "6.3.0",
     "socicon": "3.0.5",
     "tcp-ping": "0.1.1",
@@ -150,8 +149,16 @@
     "tslib": "1.10.0",
     "tslint": "5.11.0",
     "typeface-exo": "0.0.22",
-    "typescript": "3.2.4",
-    "web-animations-js": "2.2.5",
-    "zone.js": "0.8.29"
+    "typescript": "3.5.3",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.9.1"
+  },
+  "overrides": {
+    "browserslist": {
+      "caniuse-lite": "1.0.30001393"
+    },
+    "autoprefixer": {
+      "caniuse-lite": "1.0.30001393"
+    }
   }
 }
