@@ -105,7 +105,7 @@ data_symbols = {}
 for (name, sym) in symbols.items():
     if sym['type'] == 'data':
         addr = int(sym['address'], 16)
-        if 0x80400000 <= addr < 0x80420000:
+        if 0x80400000 <= addr < 0x80440000:
             addr = addr - 0x80400000 + 0x03480000
         else:
             continue
