@@ -1,8 +1,12 @@
+#include <stdbool.h>
 #include "dungeon_info.h"
-
 #include "gfx.h"
 #include "text.h"
 #include "z64.h"
+#include "trade_quests.h"
+#include "dpad.h"
+#include "item_effects.h"
+#include "save.h"
 
 typedef struct {
     uint8_t index;
@@ -14,27 +18,29 @@ typedef struct {
     };
     uint8_t skulltulas;
     char name[10];
+    uint8_t silver_rupee_puzzles_vanilla[4];
+    uint8_t silver_rupee_puzzles_mq[4];
 } dungeon_entry_t;
 
 dungeon_entry_t dungeons[] = {
-    {  8, 0, 0, 0, 0, 0x00, "Free"    },
+    {  0, 0, 0, 0, 1, 0x0F, "Deku",    {-1, -1, -1, -1}, {-1, -1, -1, -1} },
+    {  1, 0, 0, 0, 1, 0x1F, "Dodongo", {-1, -1, -1, -1}, { 0, -1, -1, -1} },
+    {  2, 0, 0, 0, 1, 0x0F, "Jabu",    {-1, -1, -1, -1}, {-1, -1, -1, -1} },
 
-    {  0, 0, 0, 0, 1, 0x0F, "Deku"    },
-    {  1, 0, 0, 0, 1, 0x1F, "Dodongo" },
-    {  2, 0, 0, 0, 1, 0x0F, "Jabu"    },
+    {  3, 1, 1, 0, 1, 0x1F, "Forest",  {-1, -1, -1, -1}, {-1, -1, -1, -1} },
+    {  4, 1, 1, 0, 1, 0x1F, "Fire",    {-1, -1, -1, -1}, {-1, -1, -1, -1} },
+    {  5, 1, 1, 0, 1, 0x1F, "Water",   {-1, -1, -1, -1}, {-1, -1, -1, -1} },
+    {  7, 1, 1, 0, 1, 0x1F, "Shadow",  { 4,  6,  7, -1}, { 4,  5,  6,  7} },
+    {  6, 1, 1, 0, 1, 0x1F, "Spirit",  {11, 14, 12, -1}, {13, 15, -1, -1} },
 
-    {  3, 1, 1, 0, 1, 0x1F, "Forest"  },
-    {  4, 1, 1, 0, 1, 0x1F, "Fire"    },
-    {  5, 1, 1, 0, 1, 0x1F, "Water"   },
-    {  7, 1, 1, 0, 1, 0x1F, "Shadow"  },
-    {  6, 1, 1, 0, 1, 0x1F, "Spirit"  },
-
-    {  8, 1, 0, 0, 1, 0x07, "BotW"    },
-    {  9, 0, 0, 0, 1, 0x07, "Ice"     },
-    { 12, 1, 0, 1, 0, 0x00, "Hideout" },
-    { 11, 1, 0, 0, 0, 0x00, "GTG"     },
-    { 13, 1, 1, 0, 0, 0x00, "Ganon"   },
+    {  8, 1, 0, 0, 1, 0x07, "BotW",    { 3, -1, -1, -1}, {-1, -1, -1, -1} },
+    {  9, 0, 0, 0, 1, 0x07, "Ice",     { 1,  2, -1, -1}, {-1, -1, -1, -1} },
+    { 12, 1, 0, 1, 0, 0x00, "Hideout", {-1, -1, -1, -1}, {-1, -1, -1, -1} },
+    { 11, 1, 0, 0, 0, 0x00, "GTG",     { 8,  9, 10, -1}, { 8,  9, 10, -1} },
+    { 13, 1, 1, 0, 0, 0x00, "Ganon",   {16, 17, 18, 21}, {18, 19, 20, -1} },
 };
+
+int dungeon_count = 13;
 
 typedef struct {
     uint8_t idx;
@@ -52,6 +58,37 @@ medal_t medals[] = {
     { 3, 0xFF, 0x82, 0x00 }, // Spirit
 };
 
+typedef struct {
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+} silver_rupee_color_data_t;
+
+silver_rupee_color_data_t silver_rupee_menu_colors[0x16][2] = {
+    {{0x00, 0x00, 0x00}, {0xFF, 0xFF, 0xFF}}, // Dodongos Cavern Staircase. Patched to use switch flag 0x1F
+    {{0x00, 0xFF, 0xFF}, {0x00, 0x00, 0x00}}, // Ice Cavern Spinning Scythe
+    {{0x00, 0x64, 0xFF}, {0x00, 0x00, 0x00}}, // Ice Cavern Push Block
+    {{0xFF, 0xFF, 0xFF}, {0x00, 0x00, 0x00}}, // Bottom of the Well Basement
+    {{0x00, 0xFF, 0x00}, {0x00, 0xFF, 0x00}}, // Shadow Temple Scythe Shortcut
+    {{0x00, 0x00, 0x00}, {0x00, 0xFF, 0xFF}}, // Shadow Temple Invisible Blades
+    {{0xC8, 0xC8, 0x00}, {0xC8, 0xC8, 0x00}}, // Shadow Temple Huge Pit
+    {{0xC8, 0x32, 0xFF}, {0xC8, 0x32, 0xFF}}, // Shadow Temple Invisible Spikes
+    {{0xC8, 0xC8, 0x00}, {0xC8, 0xC8, 0x00}}, // Gerudo Training Ground Slopes
+    {{0xFF, 0x3C, 0x00}, {0xFF, 0x3C, 0x00}}, // Gerudo Training Ground Lava
+    {{0x00, 0x64, 0xFF}, {0x00, 0x64, 0xFF}}, // Gerudo Training Ground Water
+    {{0xFF, 0x3C, 0x00}, {0x00, 0x00, 0x00}}, // Spirit Temple Child Early Torches
+    {{0x00, 0xFF, 0x00}, {0x00, 0x00, 0x00}}, // Spirit Temple Adult Boulders
+    {{0x00, 0x00, 0x00}, {0x00, 0xFF, 0xFF}}, // Spirit Temple Lobby and Lower Adult. Patched to use switch flag 0x1F
+    {{0xC8, 0xC8, 0x00}, {0x00, 0x00, 0x00}}, // Spirit Temple Sun Block
+    {{0x00, 0x00, 0x00}, {0x00, 0x64, 0xFF}}, // Spirit Temple Adult Climb
+    {{0xC8, 0xC8, 0x00}, {0x00, 0x00, 0x00}}, // Ganons Castle Spirit Trial
+    {{0x00, 0xFF, 0xFF}, {0x00, 0x00, 0x00}}, // Ganons Castle Light Trial
+    {{0xFF, 0x3C, 0x00}, {0xFF, 0x3C, 0x00}}, // Ganons Castle Fire Trial
+    {{0x00, 0x00, 0x00}, {0xC8, 0x32, 0xFF}}, // Ganons Castle Shadow Trial
+    {{0x00, 0x00, 0x00}, {0x00, 0x64, 0xFF}}, // Ganons Castle Water Trial
+    {{0x00, 0xFF, 0x00}, {0x00, 0x00, 0x00}}, // Ganons Castle Forest Trial
+};
+
 uint8_t reward_rows[] = { 0, 1, 2, 4, 5, 6, 8, 7, 3 };
 
 extern uint32_t CFG_DUNGEON_INFO_MQ_ENABLE;
@@ -59,8 +96,13 @@ extern uint32_t CFG_DUNGEON_INFO_MQ_NEED_MAP;
 extern uint32_t CFG_DUNGEON_INFO_REWARD_ENABLE;
 extern uint32_t CFG_DUNGEON_INFO_REWARD_NEED_COMPASS;
 extern uint32_t CFG_DUNGEON_INFO_REWARD_NEED_ALTAR;
+extern uint32_t CFG_DUNGEON_INFO_REWARD_SUMMARY_ENABLE;
 
 extern int8_t CFG_DUNGEON_REWARDS[14];
+extern char CFG_DUNGEON_REWARD_AREAS[9][0x17];
+extern uint8_t CFG_DUNGEON_INFO_SILVER_RUPEES;
+
+extern extended_savecontext_static_t extended_savectx;
 
 void draw_background(z64_disp_buf_t *db, int bg_left, int bg_top, int bg_width, int bg_height) {
     gDPSetCombineMode(db->p++, G_CC_PRIMITIVE, G_CC_PRIMITIVE);
@@ -76,9 +118,18 @@ void draw_background(z64_disp_buf_t *db, int bg_left, int bg_top, int bg_width, 
     gDPSetCombineMode(db->p++, G_CC_MODULATEIA_PRIM, G_CC_MODULATEIA_PRIM);
 }
 
+// skip dungeons with no keys or silver rupees
+int d_right_dungeon_idx(int i) {
+    int dungeon_idx = i + 1; // skip Free & Deku
+    if (!CFG_DUNGEON_INFO_SILVER_RUPEES || !CFG_DUNGEON_IS_MQ[DODONGO_ID]) dungeon_idx++; // skip DC
+    if (dungeon_idx >= 2) dungeon_idx++; // skip Jabu
+    if (dungeon_idx >= 9 && (!CFG_DUNGEON_INFO_SILVER_RUPEES || CFG_DUNGEON_IS_MQ[ICE_ID])) dungeon_idx++; // skip Ice
+    return dungeon_idx;
+}
+
 void draw_dungeon_info(z64_disp_buf_t *db) {
     pad_t pad_held = z64_ctxt.input[0].raw.pad;
-    int draw = CAN_DRAW_DUNGEON_INFO && (pad_held.dl || pad_held.dr || pad_held.dd);
+    int draw = CAN_DRAW_DUNGEON_INFO && !CAN_DRAW_TRADE_DPAD && (pad_held.dl || pad_held.dr || pad_held.dd || pad_held.a);
     if (!draw) {
         return;
     }
@@ -88,7 +139,326 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
     // Call setup display list
     gSPDisplayList(db->p++, &setup_db);
 
-    if (pad_held.dd) {
+    if (pad_held.a) {
+        uint16_t altar_flags = z64_file.inf_table[27];
+        int show_medals = CFG_DUNGEON_INFO_REWARD_ENABLE && (!CFG_DUNGEON_INFO_REWARD_NEED_ALTAR || (altar_flags & 1)) && CFG_DUNGEON_INFO_REWARD_SUMMARY_ENABLE;
+        int show_stones = CFG_DUNGEON_INFO_REWARD_ENABLE && (!CFG_DUNGEON_INFO_REWARD_NEED_ALTAR || (altar_flags & 2)) && CFG_DUNGEON_INFO_REWARD_SUMMARY_ENABLE;
+        int show_keys = 1;
+        int show_map_compass = 1;
+        int show_skulls = 1;
+        int show_mq = CFG_DUNGEON_INFO_MQ_ENABLE;
+
+        // Set up dimensions
+
+        int icon_size = 12;
+        int font_width = 6;
+        int font_height = 11;
+        int padding = 1;
+        int rows = 13;
+        int mq_width = show_mq ?
+            ((6 * font_width) + padding) :
+            0;
+        int silver_width = CFG_DUNGEON_INFO_SILVER_RUPEES ?
+            icon_size + (7 * font_width) + (12 * padding) :
+            0;
+        int bg_width =
+            (6 * icon_size) +
+            (11 * font_width) +
+            (8 * padding) +
+            mq_width +
+            silver_width;
+        int bg_height = (rows * icon_size) + ((rows + 1) * padding);
+        int bg_left = (Z64_SCREEN_WIDTH - bg_width) / 2;
+        int bg_top = (Z64_SCREEN_HEIGHT - bg_height) / 2;
+
+        int left = bg_left + padding;
+        int start_top = bg_top + padding;
+
+        // Draw background
+
+        gDPSetCombineMode(db->p++, G_CC_PRIMITIVE, G_CC_PRIMITIVE);
+        gDPSetPrimColor(db->p++, 0, 0, 0x00, 0x00, 0x00, 0xD0);
+        gSPTextureRectangle(db->p++,
+                bg_left<<2, bg_top<<2,
+                (bg_left + bg_width)<<2, (bg_top + bg_height)<<2,
+                0,
+                0, 0,
+                1<<10, 1<<10);
+        
+        gDPPipeSync(db->p++);
+        gDPSetCombineMode(db->p++, G_CC_MODULATEIA_PRIM, G_CC_MODULATEIA_PRIM);
+
+        // Draw medals
+
+        if (show_medals) {
+            sprite_load(db, &medals_sprite, 0, medals_sprite.tile_count);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (CFG_DUNGEON_INFO_REWARD_NEED_COMPASS && !z64_file.dungeon_items[d->index].compass) {
+                    continue;
+                }
+                int reward = CFG_DUNGEON_REWARDS[d->index];
+                if (reward < 3) continue;
+                reward -= 3;
+
+                // Medal color index was changed to hint order,
+                // moving Light from the end to the beginning.
+                // Spirit/Shadow are also swapped.
+                int reward_index;
+                if (reward < 3) {
+                    reward_index = reward + 1;
+                } else if (reward == 3) {
+                    reward_index = 5;
+                } else if (reward == 4) {
+                    reward_index = 4;
+                } else if (reward == 5) {
+                    reward_index = 0;
+                }
+                medal_t *c = &(medals[reward_index]);
+                gDPSetPrimColor(db->p++, 0, 0, c->r, c->g, c->b, 0xFF);
+
+                int top = start_top + ((icon_size + padding) * i);
+                sprite_draw(db, &medals_sprite, reward,
+                        left, top, icon_size, icon_size);
+            }
+        }
+
+        gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xFF);
+
+        // Draw stones
+
+        if (show_stones) {
+            sprite_load(db, &stones_sprite, 0, stones_sprite.tile_count);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (CFG_DUNGEON_INFO_REWARD_NEED_COMPASS && !z64_file.dungeon_items[d->index].compass) {
+                    continue;
+                }
+                int reward = CFG_DUNGEON_REWARDS[d->index];
+                if (reward < 0 || reward >= 3) continue;
+
+                int top = start_top + ((icon_size + padding) * i);
+                sprite_draw(db, &stones_sprite, reward,
+                        left, top, icon_size, icon_size);
+            }
+        }
+
+        left += icon_size + padding;
+
+        // Draw dungeon names
+
+        for (int i = 0; i < dungeon_count; i++) {
+            dungeon_entry_t *d = &(dungeons[i]);
+            int top = start_top + ((icon_size + padding) * i) + 1;
+            text_print_size(d->name, left, top, font_width);
+        }
+        text_flush_size(db, font_width, font_height, 0, 0);
+
+        left += (8 * font_width) + padding;
+
+        // Draw keys
+        
+        if (show_keys) {
+            // Draw small key counts
+
+            sprite_load(db, &quest_items_sprite, 17, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (!d->has_keys) continue;
+
+                int8_t current_keys = z64_file.dungeon_keys[d->index];
+                if (current_keys < 0) current_keys = 0;
+                if (current_keys > 9) current_keys = 9;
+
+                int8_t total_keys = z64_file.scene_flags[d->index].unk_00_ >> 0x10;
+                if (total_keys < 0) total_keys = 0;
+                if (total_keys > 9) total_keys = 9;
+
+                char count[5] = "O(O)";
+                if (current_keys > 0) count[0] = current_keys + '0';
+                if (total_keys > 0) count[2] = total_keys + '0';
+                int top = start_top + ((icon_size + padding) * i) + 1;
+                text_print_size(count, left, top, font_width);
+            }
+            text_flush_size(db, font_width, font_height, 0, 0);
+
+            left += (4 * font_width) + padding;
+
+            // Draw boss keys
+
+            sprite_load(db, &quest_items_sprite, 14, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                // Replace index 13 (Ganon's Castle) with 10 (Ganon's Tower)
+                int index = d->index == 13 ? 10 : d->index;
+
+                if (d->has_boss_key && z64_file.dungeon_items[index].boss_key) {
+                    int top = start_top + ((icon_size + padding) * i);
+                    sprite_draw(db, &quest_items_sprite, 0,
+                            left, top, icon_size, icon_size);
+                }
+            }
+
+            // Draw gerudo card
+
+            sprite_load(db, &quest_items_sprite, 10, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (d->has_card && z64_file.gerudos_card) {
+                    int top = start_top + ((icon_size + padding) * i);
+                    sprite_draw(db, &quest_items_sprite, 0,
+                            left, top, icon_size, icon_size);
+                }
+            }
+
+            left += icon_size + padding;
+        }
+
+        // Draw maps and compasses
+
+        if (show_map_compass) {
+            // Draw maps
+
+            sprite_load(db, &quest_items_sprite, 16, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (d->has_map && z64_file.dungeon_items[d->index].map) {
+                    int top = start_top + ((icon_size + padding) * i);
+                    sprite_draw(db, &quest_items_sprite, 0,
+                            left, top, icon_size, icon_size);
+                }
+            }
+
+            left += icon_size + padding;
+
+            // Draw compasses
+
+            sprite_load(db, &quest_items_sprite, 15, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (d->has_map && z64_file.dungeon_items[d->index].compass) {
+                    int top = start_top + ((icon_size + padding) * i);
+                    sprite_draw(db, &quest_items_sprite, 0,
+                            left, top, icon_size, icon_size);
+                }
+            }
+
+            left += icon_size + padding;
+        }
+
+        if (show_skulls) {
+            // Draw skulltula icon
+
+            sprite_load(db, &quest_items_sprite, 11, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (d->skulltulas && z64_file.gs_flags[d->index ^ 0x03] == d->skulltulas) {
+                    int top = start_top + ((icon_size + padding) * i);
+                    sprite_draw(db, &quest_items_sprite, 0,
+                            left, top, icon_size, icon_size);
+                }
+            }
+
+            left += icon_size + padding;
+        }
+
+        // Draw master quest dungeons
+
+        if (show_mq) {
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                if (CFG_DUNGEON_INFO_MQ_NEED_MAP && d->has_map &&
+                        !z64_file.dungeon_items[d->index].map) {
+                    continue;
+                }
+                char *str = CFG_DUNGEON_IS_MQ[d->index] ? "MQ" : "Normal";
+                int top = start_top + ((icon_size + padding) * i) + 1;
+                text_print_size(str, left, top, font_width);
+            }
+
+            left += (6 * font_width) + padding;
+        }
+        text_flush_size(db, font_width, font_height, 0, 0);
+
+        if (CFG_DUNGEON_INFO_SILVER_RUPEES) {
+            // Draw silver rupee icons
+
+            sprite_load(db, &key_rupee_clock_sprite, 1, 1);
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                bool show_silver_rupees = false;
+                uint8_t *silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
+                for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
+                    if (silver_rupee_puzzles[puzzle_idx] == (uint8_t) -1) break;
+                    uint8_t count = extended_savectx.silver_rupee_counts[silver_rupee_puzzles[puzzle_idx]];
+                    if (count > 0) {
+                        show_silver_rupees = true;
+                        break;
+                    }
+                }
+                if (show_silver_rupees) {
+                    int top = start_top + ((icon_size + padding) * i);
+                    sprite_draw(db, &key_rupee_clock_sprite, 0,
+                            left, top, icon_size, icon_size);
+                }
+            }
+
+            left += icon_size + padding;
+
+            // Draw silver rupee counts
+            sprite_load(db, &font_sprite, 16, 10); // load characters 0 through 9
+
+            for (int i = 0; i < dungeon_count; i++) {
+                dungeon_entry_t *d = &(dungeons[i]);
+                bool show_silver_rupees = false;
+                uint8_t *silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
+                for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
+                    if (silver_rupee_puzzles[puzzle_idx] == (uint8_t) -1) break;
+                    uint8_t rupee_count = extended_savectx.silver_rupee_counts[silver_rupee_puzzles[puzzle_idx]];
+                    if (rupee_count > 0) {
+                        show_silver_rupees = true;
+                        break;
+                    }
+                }
+                if (show_silver_rupees) {
+                    int top = start_top + ((icon_size + padding) * i) + 1;
+                    for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
+                        if (silver_rupee_puzzles[puzzle_idx] == (uint8_t) -1) break;
+                        silver_rupee_color_data_t color = silver_rupee_menu_colors[silver_rupee_puzzles[puzzle_idx]][CFG_DUNGEON_IS_MQ[d->index]];
+                        uint8_t rupee_count = extended_savectx.silver_rupee_counts[silver_rupee_puzzles[puzzle_idx]];
+                        int puzzle_left = left + font_width * (2 * puzzle_idx) + padding * 3 * puzzle_idx;
+                        // draw text manually instead of going through text_print/text_flush to get the right text colors
+                        gDPSetPrimColor(db->p++, 0, 0, color.r, color.g, color.b, 0xFF);
+                        if(rupee_count >= 10) {
+                            sprite_draw(db, &font_sprite, rupee_count / 10, puzzle_left, top, font_width, font_height);
+                        }
+                        int tile_index = rupee_count % 10 > 0 ? rupee_count % 10 : 0;
+                        if (tile_index == 0) {
+                            sprite_load(db, &font_sprite, 47, 1); // load letter O
+                        }
+                        sprite_draw(db, &font_sprite, tile_index, puzzle_left + font_width, top, font_width, font_height);
+                        if (tile_index == 0) {
+                            sprite_load(db, &font_sprite, 16, 10); // load numbers 0 through 9
+                        }
+                    }
+                }
+            }
+
+            gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xFF);
+        }
+
+        // Finish
+
+    } else if (pad_held.dd) {
         uint16_t altar_flags = z64_file.inf_table[27];
         int show_medals = CFG_DUNGEON_INFO_REWARD_ENABLE && (!CFG_DUNGEON_INFO_REWARD_NEED_ALTAR || (altar_flags & 1));
         int show_stones = CFG_DUNGEON_INFO_REWARD_ENABLE && (!CFG_DUNGEON_INFO_REWARD_NEED_ALTAR || (altar_flags & 2));
@@ -100,7 +470,7 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
         int rows = 9;
         int bg_width =
             (1 * icon_size) +
-            (8 * font_sprite.tile_w) +
+            (0x16 * font_sprite.tile_w) +
             (3 * padding);
         int bg_height = (rows * icon_size) + ((rows + 1) * padding);
         int bg_left = (Z64_SCREEN_WIDTH - bg_width) / 2;
@@ -141,18 +511,37 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
         // Draw dungeon names
 
         for (int i = 0; i < 9; i++) {
-            dungeon_entry_t *d = &(dungeons[i]);
-            if (i != 0 && CFG_DUNGEON_INFO_REWARD_NEED_COMPASS &&
-                    !z64_file.dungeon_items[d->index].compass) {
-                continue;
+            if (i < 3 ? show_stones : show_medals) {
+                uint8_t reward = reward_rows[i];
+                bool display_area = true;
+                switch (CFG_DUNGEON_INFO_REWARD_NEED_COMPASS) {
+                    case 1:
+                        for (int j = 0; j < 8; j++) {
+                            uint8_t dungeon_idx = dungeons[j].index;
+                            if (CFG_DUNGEON_REWARDS[dungeon_idx] == reward) {
+                                if (!z64_file.dungeon_items[dungeon_idx].compass) {
+                                    display_area = false;
+                                }
+                                break;
+                            }
+                        }
+                        break;
+                    case 2:
+                        if (i != 3) { // always display Light Medallion
+                            dungeon_entry_t *d = &(dungeons[i - (i < 3 ? 0 : 1)]); // vanilla location of the reward
+                            display_area = z64_file.dungeon_items[d->index].compass;
+                        }
+                        break;
+                }
+                if (!display_area) {
+                    continue;
+                }
+                int top = start_top + ((icon_size + padding) * i) + 1;
+                text_print(CFG_DUNGEON_REWARD_AREAS[i], left, top);
             }
-            int reward = CFG_DUNGEON_REWARDS[d->index];
-            if (i != 0 && ((reward < 3 && !show_stones) || (reward >= 3 && !show_medals))) continue;
-            int top = start_top + ((icon_size + padding) * reward_rows[reward]) + 1;
-            text_print(d->name, left, top);
         }
 
-        left += (8 * font_sprite.tile_w) + padding;
+        left += (0x16 * font_sprite.tile_w) + padding;
     } else if (pad_held.dr) {
         int show_keys = 1;
 
@@ -165,6 +554,11 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
             (1 * icon_size) +
             (12 * font_sprite.tile_w) +
             (4 * padding);
+        if (CFG_DUNGEON_INFO_SILVER_RUPEES) {
+            bg_width += icon_size + (8 * font_sprite.tile_w) + (16 * padding);
+            if (CFG_DUNGEON_IS_MQ[DODONGO_ID]) rows++;
+            if (!CFG_DUNGEON_IS_MQ[ICE_ID]) rows++;
+        }
         int bg_height = (rows * icon_size) + ((rows + 1) * padding);
         int bg_left = (Z64_SCREEN_WIDTH - bg_width) / 2;
         int bg_top = (Z64_SCREEN_HEIGHT - bg_height) / 2;
@@ -177,8 +571,8 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
 
         // Draw dungeon names
 
-        for (int i = 0; i < 9; i++) {
-            dungeon_entry_t *d = &(dungeons[i + (i > 5 ? 5 : 4)]); // skip Deku/DC/Jabu/Ice
+        for (int i = 0; i < rows; i++) {
+            dungeon_entry_t *d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
             int top = start_top + ((icon_size + padding) * i) + 1;
             text_print(d->name, left, top);
         }
@@ -187,63 +581,130 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
 
         // Draw keys
 
-        if (show_keys) {
-            // Draw small key counts
+        // Draw small key counts
 
-            sprite_load(db, &quest_items_sprite, 17, 1);
+        sprite_load(db, &quest_items_sprite, 17, 1);
 
-            for (int i = 0; i < 9; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 5 ? 5 : 4)]); // skip Deku/DC/Jabu/Ice
-                if (!d->has_keys) continue;
+        for (int i = 0; i < rows; i++) {
+            dungeon_entry_t *d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            if (!d->has_keys) continue;
 
-                int8_t current_keys = z64_file.dungeon_keys[d->index];
-                if (current_keys < 0) current_keys = 0;
-                if (current_keys > 9) current_keys = 9;
+            int8_t current_keys = z64_file.dungeon_keys[d->index];
+            if (current_keys < 0) current_keys = 0;
+            if (current_keys > 9) current_keys = 9;
 
-                int8_t total_keys = z64_file.scene_flags[d->index].unk_00_;
-                if (total_keys < 0) total_keys = 0;
-                if (total_keys > 9) total_keys = 9;
+            int8_t total_keys = z64_file.scene_flags[d->index].unk_00_ >> 0x10;
+            if (total_keys < 0) total_keys = 0;
+            if (total_keys > 9) total_keys = 9;
 
-                char count[5] = "0(0)";
-                count[0] += current_keys;
-                count[2] += total_keys;
-                int top = start_top + ((icon_size + padding) * i) + 1;
-                text_print(count, left, top);
+            char count[5] = "O(O)";
+            if (current_keys > 0) count[0] = current_keys + '0';
+            if (total_keys > 0) count[2] = total_keys + '0';
+            int top = start_top + ((icon_size + padding) * i) + 1;
+            text_print(count, left, top);
+        }
+
+        left += (4 * font_sprite.tile_w) + padding;
+
+        // Draw boss keys
+
+        sprite_load(db, &quest_items_sprite, 14, 1);
+
+        for (int i = 0; i < rows; i++) {
+            dungeon_entry_t *d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            // Replace index 13 (Ganon's Castle) with 10 (Ganon's Tower)
+            int index = d->index == 13 ? 10 : d->index;
+
+            if (d->has_boss_key && z64_file.dungeon_items[index].boss_key) {
+                int top = start_top + ((icon_size + padding) * i);
+                sprite_draw(db, &quest_items_sprite, 0,
+                        left, top, icon_size, icon_size);
             }
+        }
 
-            left += (4 * font_sprite.tile_w) + padding;
+        // Draw gerudo card
 
-            // Draw boss keys
+        sprite_load(db, &quest_items_sprite, 10, 1);
 
-            sprite_load(db, &quest_items_sprite, 14, 1);
+        for (int i = 0; i < rows; i++) {
+            dungeon_entry_t *d = &(dungeons[d_right_dungeon_idx(i)]); // skip Deku/DC/Jabu/Ice dynamically
+            if (d->has_card && z64_file.gerudos_card) {
+                int top = start_top + ((icon_size + padding) * i);
+                sprite_draw(db, &quest_items_sprite, 0,
+                        left, top, icon_size, icon_size);
+            }
+        }
 
-            for (int i = 0; i < 9; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 5 ? 5 : 4)]); // skip Deku/DC/Jabu/Ice
-                // Replace index 13 (Ganon's Castle) with 10 (Ganon's Tower)
-                int index = d->index == 13 ? 10 : d->index;
+        left += icon_size + padding;
 
-                if (d->has_boss_key && z64_file.dungeon_items[index].boss_key) {
-                    int top = start_top + ((icon_size + padding) * i);
-                    sprite_draw(db, &quest_items_sprite, 0,
-                            left, top, icon_size, icon_size);
+        if (CFG_DUNGEON_INFO_SILVER_RUPEES) {
+            // Draw silver rupee icons
+
+            sprite_load(db, &key_rupee_clock_sprite, 1, 1);
+
+            for (int i = 0; i < rows; i++) {
+                dungeon_entry_t *d = &(dungeons[d_right_dungeon_idx(i)]);
+                bool show_silver_rupees = false;
+                uint8_t *silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
+                for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
+                    if (silver_rupee_puzzles[puzzle_idx] == (uint8_t) -1) break;
+                    uint8_t count = extended_savectx.silver_rupee_counts[silver_rupee_puzzles[puzzle_idx]];
+                    if (count > 0) {
+                        show_silver_rupees = true;
+                        break;
+                    }
                 }
-            }
-
-            // Draw gerudo card
-
-            sprite_load(db, &quest_items_sprite, 10, 1);
-
-            for (int i = 0; i < 9; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 5 ? 5 : 4)]); // skip Deku/DC/Jabu/Ice
-                if (d->has_card && z64_file.gerudos_card) {
+                if (show_silver_rupees) {
                     int top = start_top + ((icon_size + padding) * i);
-                    sprite_draw(db, &quest_items_sprite, 0,
+                    sprite_draw(db, &key_rupee_clock_sprite, 0,
                             left, top, icon_size, icon_size);
                 }
             }
 
             left += icon_size + padding;
+
+            // Draw silver rupee counts
+            sprite_load(db, &font_sprite, 16, 10); // load characters 0 through 9
+
+            for (int i = 0; i < rows; i++) {
+                dungeon_entry_t *d = &(dungeons[d_right_dungeon_idx(i)]);
+                bool show_silver_rupees = false;
+                uint8_t *silver_rupee_puzzles = CFG_DUNGEON_IS_MQ[d->index] ? d->silver_rupee_puzzles_mq : d->silver_rupee_puzzles_vanilla;
+                for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
+                    if (silver_rupee_puzzles[puzzle_idx] == (uint8_t) -1) break;
+                    uint8_t count = extended_savectx.silver_rupee_counts[silver_rupee_puzzles[puzzle_idx]];
+                    if (count > 0) {
+                        show_silver_rupees = true;
+                        break;
+                    }
+                }
+                if (show_silver_rupees) {
+                    int top = start_top + ((icon_size + padding) * i) + 1;
+                    for (int puzzle_idx = 0; puzzle_idx < 4; puzzle_idx++) {
+                        if (silver_rupee_puzzles[puzzle_idx] == (uint8_t) -1) break;
+                        silver_rupee_color_data_t color = silver_rupee_menu_colors[silver_rupee_puzzles[puzzle_idx]][CFG_DUNGEON_IS_MQ[d->index]];
+                        uint8_t count = extended_savectx.silver_rupee_counts[silver_rupee_puzzles[puzzle_idx]];
+                        int puzzle_left = left + font_sprite.tile_w * (2 * puzzle_idx) + padding * 4 * puzzle_idx;
+                        // draw text manually instead of going through text_print/text_flush to get the right text colors
+                        gDPSetPrimColor(db->p++, 0, 0, color.r, color.g, color.b, 0xFF);
+                        if(count >= 10) {
+                            sprite_draw(db, &font_sprite, count / 10, puzzle_left, top, font_sprite.tile_w, font_sprite.tile_h);
+                        }
+                        int tile_index = count % 10 > 0 ? count % 10 : 0;
+                        if (tile_index == 0) {
+                            sprite_load(db, &font_sprite, 47, 1); // load letter O
+                        }
+                        sprite_draw(db, &font_sprite, tile_index, puzzle_left + font_sprite.tile_w, top, font_sprite.tile_w, font_sprite.tile_h);
+                        if (tile_index == 0) {
+                            sprite_load(db, &font_sprite, 16, 10); // load numbers 0 through 9
+                        }
+                    }
+                }
+            }
+
+            gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xFF);
         }
+
     } else { // pad_held.dl
         int show_map_compass = 1;
         int show_skulls = 1;
@@ -275,7 +736,7 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
         // Draw dungeon names
 
         for (int i = 0; i < 12; i++) {
-            dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 2 : 1)]); // skip Hideout
+            dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
             int top = start_top + ((icon_size + padding) * i) + 1;
             text_print(d->name, left, top);
         }
@@ -290,7 +751,7 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
             sprite_load(db, &quest_items_sprite, 16, 1);
 
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 2 : 1)]); // skip Hideout
+                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
                 if (d->has_map && z64_file.dungeon_items[d->index].map) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -305,7 +766,7 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
             sprite_load(db, &quest_items_sprite, 15, 1);
 
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 2 : 1)]); // skip Hideout
+                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
                 if (d->has_map && z64_file.dungeon_items[d->index].compass) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -322,7 +783,7 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
             sprite_load(db, &quest_items_sprite, 11, 1);
 
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 2 : 1)]); // skip Hideout
+                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
                 if (d->skulltulas && z64_file.gs_flags[d->index ^ 0x03] == d->skulltulas) {
                     int top = start_top + ((icon_size + padding) * i);
                     sprite_draw(db, &quest_items_sprite, 0,
@@ -337,7 +798,7 @@ void draw_dungeon_info(z64_disp_buf_t *db) {
 
         if (show_mq) {
             for (int i = 0; i < 12; i++) {
-                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 2 : 1)]); // skip Hideout
+                dungeon_entry_t *d = &(dungeons[i + (i > 9 ? 1 : 0)]); // skip Hideout
                 if (CFG_DUNGEON_INFO_MQ_NEED_MAP && d->has_map &&
                         !z64_file.dungeon_items[d->index].map) {
                     continue;
