@@ -53,8 +53,7 @@ hash_symbol_t hash_symbols[32] = {
     { 1, 19 }, // Big Magic
 };
 
-
-uint32_t cfg_file_select_hash = 0;
+extern uint8_t CFG_FILE_SELECT_HASH[5];
 
 void draw_file_select_hash(uint32_t fade_out_alpha, z64_menudata_t* menu_data) {
     z64_disp_buf_t *db = &(z64_ctxt.gfx->poly_opa);
@@ -73,10 +72,8 @@ void draw_file_select_hash(uint32_t fade_out_alpha, z64_menudata_t* menu_data) {
     gDPSetCombineMode(db->p++, G_CC_MODULATEIA_PRIM, G_CC_MODULATEIA_PRIM);
     gDPSetPrimColor(db->p++, 0, 0, 0xFF, 0xFF, 0xFF, 0xFF);
 
-    int hash = cfg_file_select_hash;
     for (int i = 0; i < icon_count; i++) {
-        int sym_index = hash & 0x1F;
-        hash >>= 5;
+        uint8_t sym_index = CFG_FILE_SELECT_HASH[i];
         hash_symbol_t *sym_desc = &(hash_symbols[sym_index]);
         sprite_t *sym_sprite = hash_sprites[sym_desc->sprite_index];
 
@@ -85,7 +82,7 @@ void draw_file_select_hash(uint32_t fade_out_alpha, z64_menudata_t* menu_data) {
 
         left += icon_size + padding;
     }
-    
+
     draw_file_message(db, menu_data);
     draw_file_icons(db, menu_data);
 
