@@ -24,6 +24,7 @@ export class FooterComponent {
 
   localVersion: string = "";
   remoteVersion: string = "";
+  branchUrl: string = "";
   hasUpdate: boolean = false;
 
   constructor(public global: GUIGlobal, private dialogService: NbDialogService) { }
@@ -39,7 +40,8 @@ export class FooterComponent {
         }
         else if (eventObj.name == "local_version_checked") {
           this.localVersion = eventObj.version;
-        }      
+          this.branchUrl = eventObj.branchUrl;
+        }
       });
 
       if (this.global.getGlobalVar("appReady")) {
@@ -63,8 +65,11 @@ export class FooterComponent {
     this.dialogService.open(ConfirmationWindow, {
       autoFocus: true, closeOnBackdropClick: true, closeOnEsc: true, hasBackdrop: true, hasScroll: false, context: { dialogHeader: "New Version Available!", dialogMessage: "You are using version " + this.localVersion + ", and the latest is version " + this.remoteVersion + ". Do you want to download the latest version now?" + ((this.remoteVersion.includes("Release")) ? "" : " (Note that you are using a development build and therefore will have to redownload and compile the source off GitHub yourself)") }
     }).onClose.subscribe(confirmed => {
-      if (confirmed)
-        (<any>window).open("https://github.com/Roman971/OoT-Randomizer/tree/Dev-R", "_blank");
+
+      if (confirmed) {
+        let link = this.remoteVersion.includes("Release") ? "https://www.ootrandomizer.com/downloads" : this.branchUrl;
+        (<any>window).open(link, "_blank");
+      }
     });
   }
 }
