@@ -1,5 +1,6 @@
-import { Component, ChangeDetectorRef, OnInit } from '@angular/core';
-import { GUIGlobal } from '../../../providers/GUIGlobal';
+import {Component, ChangeDetectorRef} from '@angular/core';
+import {GUIGlobal} from '../../../providers/GUIGlobal';
+import {NbThemeService} from "@nebular/theme";
 
 @Component({
   selector: 'ngx-header',
@@ -10,12 +11,16 @@ export class HeaderComponent {
 
   isMaximized: boolean = false;
   platform: string = (<any>window).apiPlatform;
+  isDark = false;
 
-  constructor(private cd: ChangeDetectorRef, public global: GUIGlobal) { }
+  constructor(private cd: ChangeDetectorRef,
+              public global: GUIGlobal,
+              private themeService: NbThemeService) { }
 
   ngOnInit() {
 
     if (this.global.getGlobalVar('electronAvailable')) {
+      this.initTheme();
 
       this.global.isWindowMaximized().then(res => {
         this.isMaximized = res;
@@ -54,7 +59,27 @@ export class HeaderComponent {
     this.global.maximizeWindow();
   }
 
+  switchTheme() {
+    if (this.isDark) {
+      this.themeService.changeTheme('ootr-default');
+      this.isDark = false;
+      this.global.generator_settingsMap["theme"] = 'ootr-default';
+    } else {
+      this.themeService.changeTheme('ootr-dark');
+      this.isDark = true;
+      this.global.generator_settingsMap["theme"] = 'ootr-dark';
+    }
+    this.global.saveCurrentSettingsToFile();
+  }
+
   close() {
     this.global.closeWindow();
+  }
+
+  private initTheme() {
+    if (this.global.generator_settingsMap && this.global.generator_settingsMap["theme"]) {
+      this.themeService.changeTheme(this.global.generator_settingsMap["theme"]);
+      this.isDark = 'ootr-dark' === this.global.generator_settingsMap["theme"];
+    }
   }
 }
