@@ -17,11 +17,12 @@ item_table = {
     'Deliver Letter':                                  ('Item',     True,  None, {'bottle': float('Inf')}),
     'Sell Big Poe':                                    ('Item',     True,  None, {'bottle': float('Inf')}),
     'Magic Bean':                                      ('Item',     True,  0x16, {'progressive': 10}),
-    'Skull Mask':                                      ('Item',     None,  0x17, {'trade': True}),
-    'Spooky Mask':                                     ('Item',     None,  0x18, {'trade': True}),
-    'Keaton Mask':                                     ('Item',     None,  0x1A, {'trade': True}),
-    'Bunny Hood':                                      ('Item',     None,  0x1B, {'trade': True}),
-    'Mask of Truth':                                   ('Item',     None,  0x1C, {'trade': True}),
+    'Skull Mask':                                      ('Item',     True,  0x17, {'trade': True, 'object': 0x0136}),
+    'Spooky Mask':                                     ('Item',     True,  0x18, {'trade': True, 'object': 0x0135}),
+    'Chicken':                                         ('Item',     True,  0x19, {'trade': True}),
+    'Keaton Mask':                                     ('Item',     True,  0x1A, {'trade': True, 'object': 0x0134}),
+    'Bunny Hood':                                      ('Item',     True,  0x1B, {'trade': True, 'object': 0x0137}),
+    'Mask of Truth':                                   ('Item',     True,  0x1C, {'trade': True, 'object': 0x0138}),
     'Pocket Egg':                                      ('Item',     True,  0x1D, {'trade': True}),
     'Pocket Cucco':                                    ('Item',     True,  0x1E, {'trade': True}),
     'Odd Mushroom':                                    ('Item',     True,  0x1F, {'trade': True}),
@@ -33,7 +34,7 @@ item_table = {
     'Eyedrops':                                        ('Item',     True,  0x25, {'trade': True}),
     'Claim Check':                                     ('Item',     True,  0x26, {'trade': True}),
     'Kokiri Sword':                                    ('Item',     True,  0x27, None),
-    'Giants Knife':                                    ('Item',     True,  0x28, None),
+    'Giants Knife':                                    ('Item',     None,  0x28, None),
     'Deku Shield':                                     ('Item',     None,  0x29, None),
     'Hylian Shield':                                   ('Item',     None,  0x2A, None),
     'Mirror Shield':                                   ('Item',     True,  0x2B, None),
@@ -58,12 +59,12 @@ item_table = {
     'Rupees (5)':                                      ('Item',     None,  0x4D, {'junk': 10}),
     'Rupees (20)':                                     ('Item',     None,  0x4E, {'junk': 4}),
     'Milk':                                            ('Item',     None,  0x50, None),
-    'Goron Mask':                                      ('Item',     None,  0x51, None),
-    'Zora Mask':                                       ('Item',     None,  0x52, None),
-    'Gerudo Mask':                                     ('Item',     None,  0x53, None),
+    'Goron Mask':                                      ('Item',     None,  0x51, {'trade': True, 'object': 0x0150}),
+    'Zora Mask':                                       ('Item',     None,  0x52, {'trade': True, 'object': 0x0151}),
+    'Gerudo Mask':                                     ('Item',     None,  0x53, {'trade': True, 'object': 0x0152}),
     'Rupees (50)':                                     ('Item',     None,  0x55, {'junk': 1}),
     'Rupees (200)':                                    ('Item',     None,  0x56, {'junk': 0}),
-    'Biggoron Sword':                                  ('Item',     True,  0x57, None),
+    'Biggoron Sword':                                  ('Item',     None,  0x57, None),
     'Fire Arrows':                                     ('Item',     True,  0x58, None),
     'Ice Arrows':                                      ('Item',     True,  0x59, None),
     'Light Arrows':                                    ('Item',     True,  0x5A, None),
@@ -137,7 +138,7 @@ item_table = {
     'Small Key (Gerudo Training Ground)':              ('SmallKey', True,  0xB5, {'progressive': float('Inf')}),
     'Small Key (Thieves Hideout)':              ('HideoutSmallKey', True,  0xB6, {'progressive': float('Inf')}),
     'Small Key (Ganons Castle)':                       ('SmallKey', True,  0xB7, {'progressive': float('Inf')}),
-    'Double Defense':                                  ('Item',     True,  0xB8, None),
+    'Double Defense':                                  ('Item',     None,  0xB8, None),
     'Buy Magic Bean':                                  ('Item',     True,  0x16, {'alias': ('Magic Bean', 10), 'progressive': 10}),
     'Magic Bean Pack':                                 ('Item',     True,  0xC9, {'alias': ('Magic Bean', 10), 'progressive': 10}),
     'Triforce Piece':                                  ('Item',     True,  0xCA, {'progressive': float('Inf')}),
@@ -155,6 +156,52 @@ item_table = {
     'Small Key Ring (Gerudo Training Ground)':         ('SmallKey', True,  0xD1, {'alias': ('Small Key (Gerudo Training Ground)', 10), 'progressive': float('Inf')}),
     'Small Key Ring (Thieves Hideout)':         ('HideoutSmallKey', True,  0xD2, {'alias': ('Small Key (Thieves Hideout)', 10), 'progressive': float('Inf')}),
     'Small Key Ring (Ganons Castle)':                  ('SmallKey', True,  0xD3, {'alias': ('Small Key (Ganons Castle)', 10), 'progressive': float('Inf')}),
+    
+    'Silver Rupee (Dodongos Cavern Staircase)':                 ('SilverRupee', True,  0xD4, {'progressive': 5, 'dungeon': 'Dodongos Cavern'}),
+    'Silver Rupee (Ice Cavern Spinning Blade)':                 ('SilverRupee', True,  0xD5, {'progressive': 5, 'dungeon': 'Ice Cavern'}),
+    'Silver Rupee (Ice Cavern Push Block)':                     ('SilverRupee', True,  0xD6, {'progressive': 5, 'dungeon': 'Ice Cavern'}),
+    'Silver Rupee (Bottom of the Well Basement)':               ('SilverRupee', True,  0xD7, {'progressive': 5, 'dungeon': 'Bottom of the Well'}),
+    'Silver Rupee (Shadow Temple Scythe Shortcut)':             ('SilverRupee', True,  0xD8, {'progressive': 5, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee (Shadow Temple Invisible Blades)':            ('SilverRupee', True,  0xD9, {'progressive': 10, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee (Shadow Temple Huge Pit)':                    ('SilverRupee', True,  0xDA, {'progressive': 5, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee (Shadow Temple Invisible Spikes)':            ('SilverRupee', True,  0xDB, {'progressive': 10, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee (Gerudo Training Ground Slopes)':             ('SilverRupee', True,  0xDC, {'progressive': 5, 'dungeon': 'Gerudo Training Ground'}),
+    'Silver Rupee (Gerudo Training Ground Lava)':               ('SilverRupee', True,  0xDD, {'progressive': 6, 'dungeon': 'Gerudo Training Ground'}),
+    'Silver Rupee (Gerudo Training Ground Water)':              ('SilverRupee', True,  0xDE, {'progressive': 5, 'dungeon': 'Gerudo Training Ground'}),
+    'Silver Rupee (Spirit Temple Child Early Torches)':         ('SilverRupee', True,  0xDF, {'progressive': 5, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee (Spirit Temple Adult Boulders)':              ('SilverRupee', True,  0xE0, {'progressive': 5, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee (Spirit Temple Lobby and Lower Adult)':       ('SilverRupee', True,  0xE1, {'progressive': 5, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee (Spirit Temple Sun Block)':                   ('SilverRupee', True,  0xE2, {'progressive': 5, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee (Spirit Temple Adult Climb)':                 ('SilverRupee', True,  0xE3, {'progressive': 5, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee (Ganons Castle Spirit Trial)':                ('SilverRupee', True,  0xE4, {'progressive': 5, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee (Ganons Castle Light Trial)':                 ('SilverRupee', True,  0xE5, {'progressive': 5, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee (Ganons Castle Fire Trial)':                  ('SilverRupee', True,  0xE6, {'progressive': 5, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee (Ganons Castle Shadow Trial)':                ('SilverRupee', True,  0xE7, {'progressive': 5, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee (Ganons Castle Water Trial)':                 ('SilverRupee', True,  0xE8, {'progressive': 5, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee (Ganons Castle Forest Trial)':                ('SilverRupee', True,  0xE9, {'progressive': 5, 'dungeon': 'Ganons Castle'}),
+
+    'Silver Rupee Pouch (Dodongos Cavern Staircase)':           ('SilverRupee', True,  0xEA, {'alias': ('Silver Rupee (Dodongos Cavern Staircase)', 10), 'progressive': 1, 'dungeon': 'Dodongos Cavern'}),
+    'Silver Rupee Pouch (Ice Cavern Spinning Blade)':           ('SilverRupee', True,  0xEB, {'alias': ('Silver Rupee (Ice Cavern Spinning Scythe)', 10), 'progressive': 1, 'dungeon': 'Ice Cavern'}),
+    'Silver Rupee Pouch (Ice Cavern Push Block)':               ('SilverRupee', True,  0xEC, {'alias': ('Silver Rupee (Ice Cavern Push Block)', 10), 'progressive': 1, 'dungeon': 'Ice Cavern'}),
+    'Silver Rupee Pouch (Bottom of the Well Basement)':         ('SilverRupee', True,  0xED, {'alias': ('Silver Rupee (Bottom of the Well Basement)', 10), 'progressive': 1, 'dungeon': 'Bottom of the Well'}),
+    'Silver Rupee Pouch (Shadow Temple Scythe Shortcut)':       ('SilverRupee', True,  0xEE, {'alias': ('Silver Rupee (Shadow Temple Scythe Shortcut)', 10), 'progressive': 1, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee Pouch (Shadow Temple Invisible Blades)':      ('SilverRupee', True,  0xEF, {'alias': ('Silver Rupee (Shadow Temple Invisible Blades)', 10), 'progressive': 1, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee Pouch (Shadow Temple Huge Pit)':              ('SilverRupee', True,  0xF0, {'alias': ('Silver Rupee (Shadow Temple Huge Pit)', 10), 'progressive': 1, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee Pouch (Shadow Temple Invidible Spikes)':      ('SilverRupee', True,  0xF1, {'alias': ('Silver Rupee (Shadow Temple Invidible Spikes)', 10), 'progressive': 1, 'dungeon': 'Shadow Temple'}),
+    'Silver Rupee Pouch (Gerudo Training Ground Slopes)':       ('SilverRupee', True,  0xF2, {'alias': ('Silver Rupee (Gerudo Training Ground Slopes)', 10), 'progressive': 1, 'dungeon': 'Gerudo Training Ground'}),
+    'Silver Rupee Pouch (Gerudo Training Ground Lava)':         ('SilverRupee', True,  0xF3, {'alias': ('Silver Rupee (Gerudo Training Ground Lava)', 10), 'progressive': 1, 'dungeon': 'Gerudo Training Ground'}),
+    'Silver Rupee Pouch (Gerudo Training Ground Water)':        ('SilverRupee', True,  0xF4, {'alias': ('Silver Rupee (Gerudo Training Ground Water)', 10), 'progressive': 1, 'dungeon': 'Gerudo Training Ground'}),
+    'Silver Rupee Pouch (Spirit Temple Child Early Torches)':   ('SilverRupee', True,  0xF5, {'alias': ('Silver Rupee (Spirit Temple Child Early Torches)', 10), 'progressive': 1, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee Pouch (Spirit Temple Adult Boulders)':        ('SilverRupee', True,  0xF6, {'alias': ('Silver Rupee (Spirit Temple Adult Boulders)', 10), 'progressive': 1, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee Pouch (Spirit Temple Lobby and Lower Adult)': ('SilverRupee', True,  0xF7, {'alias': ('Silver Rupee (Spirit Temple Lobby and Lower Adult)', 10), 'progressive': 1, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee Pouch (Spirit Temple Sun Block)':             ('SilverRupee', True,  0xF8, {'alias': ('Silver Rupee (Spirit Temple Sun Block)', 10), 'progressive': 1, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee Pouch (Spirit Temple Adult Climb)':           ('SilverRupee', True,  0xF9, {'alias': ('Silver Rupee (Spirit Temple Adult Climb)', 10), 'progressive': 1, 'dungeon': 'Spirit Temple'}),
+    'Silver Rupee Pouch (Ganons Castle Spirit Trial)':          ('SilverRupee', True,  0xFA, {'alias': ('Silver Rupee (Ganons Castle Spirit Trial)', 10), 'progressive': 1, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee Pouch (Ganons Castle Light Trial)':           ('SilverRupee', True,  0xFB, {'alias': ('Silver Rupee (Ganons Castle Light Trial)', 10), 'progressive': 1, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee Pouch (Ganons Castle Fire Trial)':            ('SilverRupee', True,  0xFC, {'alias': ('Silver Rupee (Ganons Castle Fire Trial)', 10), 'progressive': 1, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee Pouch (Ganons Castle Shadow Trial)':          ('SilverRupee', True,  0xFD, {'alias': ('Silver Rupee (Ganons Castle Shadow Trial)', 10), 'progressive': 1, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee Pouch (Ganons Castle Water Trial)':           ('SilverRupee', True,  0xFE, {'alias': ('Silver Rupee (Ganons Castle Water Trial)', 10), 'progressive': 1, 'dungeon': 'Ganons Castle'}),
+    'Silver Rupee Pouch (Ganons Castle Forest Trial)':          ('SilverRupee', True,  0xFF, {'alias': ('Silver Rupee (Ganons Castle Forest Trial)', 10), 'progressive': 1, 'dungeon': 'Ganons Castle'}),
 
     'Silver Rupee (Dodongos Cavern Staircase)':                 ('SilverRupee', True,  0xD4, {'progressive': 5}),
     'Silver Rupee (Ice Cavern Spinning Scythe)':                ('SilverRupee', True,  0xD5, {'progressive': 5}),
