@@ -31,7 +31,7 @@ inventory = dict(chain(
     _entry('dins_fire', 'Dins Fire', guitext="Din's Fire"),
     _entry('slingshot', available=3, ammo={'Deku Seeds': (30, 40, 50)}),
     _entry('ocarina', available=2),
-    _entry('bombchus', ammo={'Bombchus': (19,)}), # start with additional bombchus
+    _entry('bombchus', ammo={'Bombchus': (20,)}), # start with additional bombchus
     _entry('hookshot', 'Progressive Hookshot', available=2),
     _entry('ice_arrow', 'Ice Arrows'),
     _entry('farores_wind', 'Farores Wind', guitext="Farore's Wind"),
@@ -43,6 +43,28 @@ inventory = dict(chain(
     _entry('nayrus_love', 'Nayrus Love', guitext="Nayru's Love"),
     _entry('bottle', available=3, special=True),
     _entry('letter', 'Rutos Letter', guitext="Ruto's Letter", special=True),
+    _entry("pocket_egg",   "Pocket Egg", guitext="Pocket Egg"),
+    _entry("pocket_cucco", "Pocket Cucco", guitext="Pocket Cucco"),
+    _entry("cojiro",       "Cojiro", guitext="Cojiro"),
+    _entry("odd_mushroom", "Odd Mushroom", guitext="Odd Mushroom"),
+    _entry("odd_potion",   "Odd Potion", guitext="Odd Potion"),
+    _entry("poachers_saw", "Poachers Saw", guitext="Poacher's Saw"),
+    _entry("broken_sword", "Broken Sword", guitext="Broken Sword"),
+    _entry("prescription", "Prescription", guitext="Prescription"),
+    _entry("eyeball_frog", "Eyeball Frog", guitext="Eyeball Frog"),
+    _entry("eyedrops",     "Eyedrops", guitext="Eyedrops"),
+    _entry("claim_check",  "Claim Check", guitext="Claim Check"),
+    _entry("weird_egg",    "Weird Egg", guitext="Weird Egg"),
+    _entry("chicken",      "Chicken", guitext="Chicken"),
+    _entry("zeldas_letter","Zeldas Letter", guitext="Zelda's Letter"),
+    _entry("keaton_mask",  "Keaton Mask", guitext="Keaton Mask"),
+    _entry("skull_mask",   "Skull Mask", guitext="Skull Mask"),
+    _entry("spooky_mask",  "Spooky Mask", guitext="Spooky Mask"),
+    _entry("bunny_hood",   "Bunny Hood", guitext="Bunny Hood"),
+    _entry("goron_mask",   "Goron Mask", guitext="Goron Mask"),
+    _entry("zora_mask",    "Zora Mask", guitext="Zora Mask"),
+    _entry("gerudo_mask",  "Gerudo Mask", guitext="Gerudo Mask"),
+    _entry("mask_of_truth","Mask of Truth", guitext="Mask of Truth"),
 ))
 
 songs = dict(chain(
